Fix merge conflicts.
