+++ pytket_qir/parser.py
@@ -17,36 +17,22 @@ This module contains all functionality to parse and generate QIR files
 to and from pytket circuits.
 """
 
-from enum import Enum
-from functools import partial
 import inspect
 import os
 import re
-from typing import cast, Callable, Dict, List, Optional, Sequence, Tuple, Union
+from typing import cast, Callable, Dict, List, Optional, Tuple, Union
 
 
-from pytket import Circuit, OpType, Bit, Qubit  # type: ignore
-from pytket.qasm.qasm import _retrieve_registers  # type: ignore
+from pytket import Circuit, OpType  # type: ignore
 from pytket.wasm import WasmFileHandler  # type: ignore
 from pytket.circuit import (  # type: ignore
     BitRegister,
     CircBox,
-    ClassicalExpBox,
-    Command,
-    Conditional,
     Op,
-    WASMOp,
 )
 from pytket.circuit.logic_exp import (  # type: ignore
-    BitWiseOp,
     RegAdd,
     RegAnd,
-    RegEq,
-    RegGeq,
-    RegGt,
-    RegLeq,
-    RegLt,
-    RegNeq,
     RegSub,
     RegMul,
     RegOr,
@@ -72,25 +58,21 @@ from pyqir.parser import (  # type: ignore
     QirInstr,
 )
 from pyqir.parser._native import PyQirInstruction  # type: ignore
-from pyqir.generator import SimpleModule, BasicQisBuilder, IntPredicate, ir_to_bitcode, types  # type: ignore
-from pyqir.generator._native import Value  # type: ignore
+from pyqir.generator import ir_to_bitcode, types  # type: ignore
 from pyqir.parser._native import PyQirInstruction  # type: ignore
 from pyqir.parser._parser import QirIntConstant, QirICmpInstr, QirCallInstr, QirOpInstr, QirOperand  # type: ignore
-from pyqir.generator import SimpleModule, BasicQisBuilder, ir_to_bitcode, types  # type: ignore
-from pyqir.generator.types import Qubit   # type: ignore
+from pyqir.generator import ir_to_bitcode, types  # type: ignore
 
 
 from pytket_qir.gatesets.base import (
     CustomGateSet,
-    CustomQirGate,
     OpNat,
     OpName,
     OpSpec,
     QirGate,
 )
 
-from pytket_qir.gatesets.pyqir.pyqir import PYQIR_GATES, _TK_TO_PYQIR  # type: ignore
-from pytket_qir.utils.utils import QIRFormat  # type: ignore
+from pytket_qir.gatesets.pyqir.pyqir import PYQIR_GATES  # type: ignore
 
 
 _PYQIR_TO_TK_CLOPS: Dict[str, Union[type, Dict[str, Callable]]] = {
@@ -112,31 +94,6 @@ _PYQIR_TO_TK_CLOPS: Dict[str, Union[type, Dict[str, Callable]]] = {
     "is_or": RegOr,
     "is_xor": RegXor,
 }
-
-_TK_CLOPS_TO_PYQIR: Dict = {
-    RegAnd: lambda b: b.and_,
-    RegOr: lambda b: b.or_,
-    RegXor: lambda b: b.xor,
-    RegAdd: lambda b: b.add,
-    RegSub: lambda b: b.sub,
-    RegMul: lambda b: b.mul,
-    RegLsh: lambda b: b.shl,
-    RegRsh: lambda b: b.lshr,
-    RegEq: lambda b: partial(b.icmp, IntPredicate.EQ),
-    RegNeq: lambda b: partial(b.icmp, IntPredicate.NE),
-    RegGt: lambda b: partial(b.icmp, IntPredicate.UGT),
-    RegGeq: lambda b: partial(b.icmp, IntPredicate.UGE),
-    RegLt: lambda b: partial(b.icmp, IntPredicate.ULT),
-    RegLeq: lambda b: partial(b.icmp, IntPredicate.ULE),
-}
-
-
-class CommandUnsupportedError(Exception):
-    pass
-
-
-class WASMError(Exception):
-    pass
 
 
 class QirParser:
@@ -411,346 +368,3 @@ def circuit_from_qir(
             o.write(ir_to_bitcode(data))
         input_file_str = output_bc_file
     return QirParser(input_file_str, gateset, wasm_handler, wasm_int_type).circuit
-
-
-class QIRUnsupportedError(Exception):
-    pass
-
-
-class Module:
-    """
-    PyQir module extension to account for custom defined input gate set
-    and calls to WASM files.
-    """
-
-    def __init__(
-        self,
-        name: str,
-        num_qubits: int,
-        num_results: int,
-        gateset: Optional[CustomGateSet] = None,
-        wasm_handler: Optional[WasmFileHandler] = None,
-    ) -> None:
-        self.module = SimpleModule(name, num_qubits, num_results)
-        self.builder = self.module.builder
-        self.qis = BasicQisBuilder(self.builder)
-        self.gateset = gateset if gateset is not None else PYQIR_GATES
-        self.wasm_handler = wasm_handler
-
-    @property
-    def gateset(self):
-        """A getter for the gateset."""
-        assert self._gateset is not None
-        return self._gateset
-
-    @gateset.setter
-    def gateset(self, new_gateset):
-        self._gateset = new_gateset
-        for v in self._gateset.gateset.values():
-            self.__setattr__(
-                v.opname.value,
-                self.module.add_external_function(
-                    self._gateset.template.substitute(
-                        opnat=v.opnat.value,
-                        opname=v.opname.value,
-                        opspec=v.opspec.value,
-                    ),
-                    types.Function(v.function_signature, v.return_type),
-                ),
-            )
-
-
-def _get_optype_and_params(op: Op) -> Tuple[OpType, Sequence[float]]:
-    optype = op.type
-    params: List = []
-    if optype == OpType.ExplicitPredicate:
-        if op.get_name() == "AND":
-            optype = BitWiseOp.AND
-        elif op.get_name() == "OR":
-            optype = BitWiseOp.OR
-        elif op.get_name() == "XOR":
-            optype = BitWiseOp.XOR
-    else:
-        params = op.params
-        if optype == OpType.TK1:
-            params = [op.params[1], op.params[0] - 0.5, op.params[2] + 0.5]
-    return (optype, params)
-
-
-def _to_qis_qubits(qubits: List[Qubit], mod: SimpleModule) -> Sequence[Qubit]:
-    return [mod.qubits[qubit.index[0]] for qubit in qubits]
-
-
-def _to_qis_results(
-    bits: List[Bit], mod: SimpleModule
-) -> Optional[Union[Value, float]]:
-    if bits:
-        return mod.results[bits[0].index[0]]
-    return None
-
-
-def _to_qis_bits(
-    args: List[Bit], mod: SimpleModule
-) -> Sequence[Union[Value, bool, int, float]]:
-    if args:
-        return [mod.results[bit.index[0]] for bit in args[:-1]]
-    return []
-
-
-def _reg2ssa_var(bit_reg: BitRegister, module: Module, int_size: int) -> Value:
-    # A utility function to convert from a pytket
-    #  BitRegister to an SSA variable via pyqir types.
-    bit_reg_list = list(bit_reg)
-    reg2var = module.module.add_external_function(
-        "reg2var",
-        types.Function([types.BOOL] * int_size, types.Int(int_size)),
-    )
-    if bit_reg.size <= int_size:  # Widening by zero-padding.
-        bool_reg = list(map(bool, bit_reg_list)) + [False] * (int_size - bit_reg.size)
-    else:  # Narrowing by truncation.
-        bool_reg = list(map(bool, bit_reg_list[:int_size]))
-    return cast(Value, module.builder.call(reg2var, [*bool_reg]))
-
-
-class QIRGenerator:
-    """A generator class to produce a QIR file from a pytket circuit."""
-
-    def __init__(
-        self,
-        circuit: Circuit,
-        module: Module,
-        wasm_int_type: types.Int = types.Int(32),
-    ) -> None:
-        self.circuit = circuit
-        self.cregs = _retrieve_registers(self.circuit.bits, BitRegister)
-        self.wasm_int_type = wasm_int_type
-        self.module = self.circuit_to_module(circuit, module)
-
-    def _get_c_regs_from_com(self, command: Command) -> Tuple[List[str], List[str]]:
-        op = command.op
-        args = command.args
-        inputs: List[str] = []
-        outputs: List[str] = []
-
-        if isinstance(op, WASMOp):
-            for reglist, sizes in [
-                (inputs, op.input_widths),
-                (outputs, op.output_widths),
-            ]:
-                for in_width in sizes:
-                    assert in_width > 0
-                    com_bits = args[:in_width]
-                    args = args[in_width:]
-                    regname = com_bits[0].reg_name
-                    if com_bits != list(self.cregs[regname]):
-                        CommandUnsupportedError(
-                            "Command ops must act on entire registers."
-                        )
-                    reglist.append(regname)
-        elif isinstance(op, ClassicalExpBox):
-            for reglist, sizes in [
-                (
-                    inputs,
-                    list(
-                        map(
-                            lambda obj: obj.size
-                            if isinstance(obj, BitRegister)
-                            else None,
-                            op.get_exp().args,
-                        )
-                    ),
-                ),
-                (outputs, [op.get_n_o()]),
-            ]:
-                if not sizes:
-                    ValueError(
-                        "Command op input or output registers have empty widths."
-                    )
-                for in_width in sizes:
-                    com_bits = args[:in_width]
-                    args = args[in_width:]
-                    regname = com_bits[0].reg_name
-                    if com_bits != list(self.cregs[regname]):
-                        CommandUnsupportedError(
-                            "Command ops must act on entire registers."
-                        )
-                    reglist.append(regname)
-        return inputs, outputs
-
-    def circuit_to_module(
-        self,
-        circ: Circuit,
-        module: Module,
-    ) -> Module:
-        """A method to generate a QIR string from a pytket circuit."""
-        wasm_int_size = self.wasm_int_type.width
-        for command in circ:
-            op = command.op
-            if isinstance(op, Conditional):
-                conditional_circuit = op.op.get_circuit()
-                condition_bit_index = command.args[0].index[0]
-
-                def condition_one_block():
-                    """
-                    Populate recursively the module with the contents of the conditional
-                    sub-circuit when the condition is True.
-                    """
-                    if op.value == 1:
-                        self.circuit_to_module(conditional_circuit, module)
-
-                def condition_zero_block():
-                    """
-                    Populate recursively the module with the contents of the conditional
-                    sub-circuit when the condition is False.
-                    """
-                    if op.value == 0:
-                        self.circuit_to_module(conditional_circuit, module)
-
-                module.qis.if_result(
-                    module.module.results[condition_bit_index],
-                    one=lambda: condition_one_block(),
-                    zero=lambda: condition_zero_block(),
-                )
-            elif isinstance(op, WASMOp):
-                inputs, _ = self._get_c_regs_from_com(command)
-                try:
-                    bit_reg = circ.get_c_register(inputs[0])
-                    input_type_list = [types.Int(wasm_int_size)]
-                except IndexError:
-                    input_type_list = []
-
-                # Need to create a singleton enum to hold the WASM function name.
-                WasmName = Enum("WasmName", [("WASM", op.func_name)])
-
-                # Update datastructures with WASM function name and
-                # appropriate definition.
-
-                # Update translation dict.
-                _TK_TO_PYQIR[OpType.WASM] = QirGate(
-                    opnat=OpNat.HYBRID, opname=WasmName.WASM, opspec=OpSpec.BODY
-                )
-
-                # Update gateset.
-                gateset = PYQIR_GATES.gateset
-                gateset["wasm"] = CustomQirGate(
-                    opnat=OpNat.HYBRID,
-                    opname=WasmName.WASM,
-                    opspec=OpSpec.BODY,
-                    function_signature=input_type_list,
-                    return_type=types.Int(wasm_int_size),
-                )
-
-                # Update gateset in module.
-                module.gateset = PYQIR_GATES
-
-                # Create an ssa variable if there is an input to the WASMOp.
-                if len(input_type_list) == 0:
-                    ssa_args = []
-                else:
-                    ssa_args = [_reg2ssa_var(bit_reg, module, wasm_int_size)]
-
-                gate = module.gateset.tk_to_gateset(op.type)
-                get_gate = getattr(module, gate.opname.value)
-                module.builder.call(get_gate, ssa_args)
-            elif isinstance(op, ClassicalExpBox):
-                inputs, _ = self._get_c_regs_from_com(command)
-
-                ssa_vars: List = []
-                for inp in inputs:
-                    bit_reg = circ.get_c_register(inp)
-                    ssa_vars.append(_reg2ssa_var(bit_reg, module, 64))
-
-                _TK_CLOPS_TO_PYQIR[type(op.get_exp())](module.builder)(*ssa_vars)
-            else:
-                optype, params = _get_optype_and_params(op)
-                qubits = _to_qis_qubits(command.qubits, module.module)
-                results = _to_qis_results(command.bits, module.module)
-                if module.gateset.name == "PyQir":
-                    pyqir_gate = module.gateset.tk_to_gateset(optype)
-                    if not pyqir_gate.opspec == OpSpec.BODY:
-                        opname = pyqir_gate.opname.value + "_" + pyqir_gate.opspec.value
-                        get_gate = getattr(module.qis, opname)
-                    else:
-                        get_gate = getattr(module.qis, pyqir_gate.opname.value)
-                    if params:
-                        get_gate(*params, *qubits)
-                    elif results:
-                        get_gate(*qubits, results)
-                    else:
-                        get_gate(*qubits)
-                else:
-                    bits: Optional[Sequence[Union[Value, bool, int, float]]] = None
-                    if type(optype) == BitWiseOp:
-                        bits = _to_qis_bits(command.args, module.module)
-                    gate = module.gateset.tk_to_gateset(optype)
-                    get_gate = getattr(module, gate.opname.value)
-                    if bits:
-                        module.builder.call(get_gate, bits)
-                    elif params:
-                        module.builder.call(get_gate, [*params, *qubits])
-                    elif results:
-                        module.builder.call(get_gate, [*qubits, results])
-                    else:
-                        module.builder.call(get_gate, qubits)
-        return module
-
-
-def circuit_to_qir(
-    circ: Circuit,
-    gateset: Optional[CustomGateSet] = None,
-    wasm_path: Optional[Union[str, os.PathLike]] = None,
-    wasm_int_type: Optional[types.Int] = types.Int(32),
-    qir_format: Optional[QIRFormat] = QIRFormat.BITCODE,
-) -> Union[str, bytes]:
-    """Return a pytket circuit as QIR."""
-    wasm_handler = None
-    module_name = "Generated from {} pytket circuit".format(
-        circ.name if circ.name is not None else "input"
-    )
-    if wasm_path is not None:
-        try:
-            wasm_handler = WasmFileHandler(str(wasm_path))
-            wasm_file_name = os.path.basename(str(wasm_path))
-            module_name = module_name + " and {} file.".format(wasm_file_name)
-        except ValueError as ve:
-            raise ve
-    module = Module(
-        name=module_name,
-        num_qubits=circ.n_qubits,
-        num_results=len(circ.bits),
-        gateset=gateset,
-        wasm_handler=wasm_handler,
-    )
-    populated_module = QIRGenerator(circ, module, wasm_int_type).module
-    if qir_format == QIRFormat.BITCODE:
-        return populated_module.module.bitcode()
-    else:
-        return populated_module.module.ir()
-
-
-def write_qir_file(
-    circ: Circuit,
-    file_name: str,
-    gateset: Optional[CustomGateSet] = None,
-    wasm_path: Optional[Union[str, os.PathLike]] = None,
-    wasm_int_type: Optional[types.Int] = types.Int(32),
-) -> None:
-    """A method to generate a qir file from a tket circuit."""
-    _, ext = os.path.splitext(os.path.basename(file_name))
-    if ext == ".bc":
-        qir_format = QIRFormat.BITCODE
-        file_param = "wb"
-    elif ext == ".ll":
-        qir_format = QIRFormat.IR
-        file_param = "w"
-    else:
-        raise ValueError("The file extension must either be '.ll' or '.bc'.")
-    qir = circuit_to_qir(
-        circ=circ,
-        gateset=gateset,
-        wasm_path=wasm_path,
-        wasm_int_type=wasm_int_type,
-        qir_format=qir_format,
-    )
-    with open(file_name, file_param) as out:
-        out.write(qir)
